lab-automation-testing/automation-testing/requirements.txt
# Core Testing Framework
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-xdist==3.5.0
pytest-mock==3.12.0
pytest-timeout==2.2.0

# HTTP Testing
httpx==0.25.2
requests==2.31.0

# Test Data Generation
faker==20.1.0
factory-boy==3.3.0

# Load Testing
locust==2.20.0

# Reporting
allure-pytest==2.13.2
pytest-html==4.1.1

# Code Quality
black==23.12.1
flake8==7.0.0
mypy==1.7.1
pylint==3.0.3

# FastAPI Testing
fastapi==0.108.0
uvicorn[standard]==0.25.0
sqlalchemy==2.0.23

# Django Testing
django==5.0
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database Testing
pytest-django==4.7.0
databases==0.8.0

# Mocking & Fixtures
responses==0.24.1
freezegun==1.4.0

# Performance Monitoring
pytest-benchmark==4.0.0

# Security Testing
bandit==1.7.6
safety==2.3.5
